TWINE: replaced magic numbers
