+++ engines/twine/holomap.cpp
@@ -134,7 +134,7 @@ void Holomap::processHolomap() {
 	drawHolomapTitle(320, 25);
 	_engine->_renderer->setCameraPosition(320, 190, 128, 1024, 1024);
 
-	_engine->_text->initTextBank(2);
+	_engine->_text->initTextBank(TextBankId::Inventory_Intro_and_Holomap);
 	_engine->_text->setFontCrossColor(9);
 
 	// TODO
